唤醒后转动逻辑

+++ app/src/main/java/com/geeknewbee/doraemon/input/ISoundInputDevice.java
@@ -9,7 +9,7 @@ public interface ISoundInputDevice {
 
     void setAngle(int angle);
 
-    void onWakeUp(int angle);
+    void onWakeUp(double angle);
 
     void start();
 }
